Refactor teacher dashboard to replace legacy account tabs
